+++ src/memline.c
@@ -943,7 +943,7 @@ errorret:
 		return IObuff;
 	}
 
-#if defined(KANJI) && defined(NT) && defined(SYNTAX)
+#ifdef USE_SYNTAX
 	/* The caller is about to write into the line it is being handed. */
 	if (will_change)
 		syn_changed(buf, lnum);
@@ -1038,7 +1038,7 @@ ml_append_int(BUF *buf, linenr_t lnum, char_u *line, colnr_t len, int newfile)
 	if (lnum > buf->b_ml.ml_line_count)	/* lnum out of range */
 		return FAIL;
 
-#if defined(KANJI) && defined(NT) && defined(SYNTAX)
+#ifdef USE_SYNTAX
 	syn_changed(buf, lnum);
 #endif
 	if (lowest_marked && lowest_marked > lnum)
@@ -1550,7 +1550,7 @@ ml_replace(linenr_t lnum, char_u *line, int copy)
 	if (line == NULL)			/* just checking... */
 		return FAIL;
 
-#if defined(KANJI) && defined(NT) && defined(SYNTAX)
+#ifdef USE_SYNTAX
 	syn_changed(curbuf, lnum);
 #endif
 	/*
@@ -1621,7 +1621,7 @@ ml_delete_int(BUF *buf, linenr_t lnum)
 	if (lnum < 1 || lnum > buf->b_ml.ml_line_count)
 		return FAIL;
 
-#if defined(KANJI) && defined(NT) && defined(SYNTAX)
+#ifdef USE_SYNTAX
 	syn_changed(buf, lnum);
 #endif
 	if (lowest_marked && lowest_marked > lnum)